+++ src/Labels/LabelHooks.hpp
@@ -1,3 +1,5 @@
+#ifndef GEODE_IS_ANDROID
+
 #pragma once
 
 #include <Geode/Geode.hpp>
@@ -31,3 +33,5 @@ class $modify (LabelPlayLayer, PlayLayer)
 {
     virtual void destroyPlayer(PlayerObject* p0, GameObject* p1);
 };
+
+#endif